Star Mob! (A* Search Algorithm)

+++ Rain/src/com/aadi/rain/entity/Entity.java
@@ -8,7 +8,7 @@ import com.aadi.rain.level.Level;
 
 public class Entity {
 
-	protected double x, y;
+	protected int x, y;
 	protected Sprite sprite;
 	private boolean removed = false;
 	protected Level level;
@@ -26,11 +26,11 @@ public class Entity {
 		removed = true;
 	}
 
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
